new lib runScheduledAction
