+++ lib/runScheduledAction.ts
@@ -1,0 +1,29 @@
+import { ScheduledAction } from "./supabase/scheduled_actions/getScheduledActions";
+
+export async function runScheduledAction(
+  action: ScheduledAction
+): Promise<void> {
+  const messages = [
+    {
+      role: "user",
+      content: action.prompt,
+    },
+  ];
+  const body = {
+    messages,
+    artistId: action.artist_account_id,
+    accountId: action.account_id,
+  };
+  try {
+    await fetch("https://chat.recoupable.com/api/chat", {
+      method: "POST",
+      headers: { "Content-Type": "application/json" },
+      body: JSON.stringify(body),
+    });
+  } catch (err) {
+    console.error(
+      `[runScheduledAction] Failed to call chat API for action ${action.id}:`,
+      err
+    );
+  }
+}
